+++ src/layers/core/path-layer/path-layer.js
@@ -52,8 +52,8 @@ const isClosed = path => {
 export default class PathLayer extends Layer {
   getShaders() {
     return enable64bitSupport(this.props) ?
-      {vs: vs64, fs, modules: ['project64']} :
-      {vs, fs}; // 'project' module added by default.
+      {vs: vs64, fs, modules: ['project64', 'picking']} :
+      {vs, fs, modules: ['picking']}; // 'project' module added by default.
   }
 
   initializeState() {