+++ src/layers/core/scatterplot-layer/scatterplot-layer-fragment.glsl.js
@@ -33,10 +33,15 @@ void main(void) {
 
   float distToCenter = length(unitPosition);
 
-  if (distToCenter <= 1.0 && distToCenter >= innerUnitRadius) {
-    gl_FragColor = vColor;
-  } else {
+  if (distToCenter > 1.0 || distToCenter < innerUnitRadius) {
     discard;
   }
+  gl_FragColor = vColor;
+
+  // use highlight color if this fragment belongs to the selected object.
+  gl_FragColor = picking_filterHighlightColor(gl_FragColor);
+
+  // use picking color if rendering to picking FBO.
+  gl_FragColor = picking_filterPickingColor(gl_FragColor);
 }
 `;