+++ src/shaderlib/index.js
@@ -23,13 +23,15 @@ import {fp32, fp64} from 'luma.gl';
 import project from '../shaderlib/project/project';
 import project64 from '../shaderlib/project64/project64';
 import lighting from '../shaderlib/lighting/lighting';
+import picking from '../shaderlib/picking/picking';
 
 import {registerShaderModules, setDefaultShaderModules} from 'luma.gl';
 
 registerShaderModules([
   fp32, fp64,
   project, project64,
-  lighting
+  lighting,
+  picking
 ]);
 
 setDefaultShaderModules([project]);